+++ src/pages/_app.tsx
@@ -32,11 +32,15 @@ Router.events.on('routeChangeError', () => {
   NProgress.done()
 })
 
-// axios BASE URL
+// axios default baseUrl
 axios.defaults.baseURL =
   process.env.NODE_ENV === 'development'
     ? 'http://localhost:3000/api/v1'
     : 'https://trishop.vercel.app/api/v1'
+
+// axios default validateStatus
+axios.defaults.validateStatus = (status) =>
+  (status >= 200 && status < 300) || (status >= 400 && status < 500) // Resolve only if the status code is 200 more and less than 500
 
 export default function MyApp({ Component, pageProps }: AppProps): JSX.Element {
   return (