+++ CHANGELOG.md
@@ -2,12 +2,12 @@
 
 ## 14 Juni 2021
 
-- Menambahkan bagian Change Log dan Code of Conduct pada `README.md`
+- Mengupdate `README.md`
 - Menambahkan `CHANGELOG.md`
 - Menambahkan `CODE_OF_CONDUCT.md`
 - Membuat grup komunitas Roompy di [Discord](https://discord.gg/W9gPJ6kUPY)
+- Mengubah penyimpanan user state yang awalnya dari cookies ke local storage `useLocalStorage('user')`
 
 ## 13 Juni 2021
 
 - Initial release v0.2.0
-- Menambahkan bagian Contributions pada `README.md`